+++ src/containers/SldEditor/sldEditor.css
@@ -1,7 +1,7 @@
 .container {
   display: flex;
   width: 100%;
-  height: 100%;
+  height: calc(100vh - 90px);
   overflow: hidden;
   font-size: 1rem;
   margin: 0 auto;
