+++ src/pages/SignUp.js
@@ -2,15 +2,18 @@ import React, {Fragment, useState} from "react";
 import {connect} from "react-redux";
 import {FormattedMessage} from "react-intl";
 import {Link} from "react-router-dom";
+import {useFirestore} from "react-redux-firebase";
+import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth";
+
+import Header from "../containers/Header/Header";
+import {writeDbUser} from "../lib/writeDb";
+import firebase from "../config/fbConfig";
+import "./sign.css";
+
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
 import TextField from "@material-ui/core/TextField";
-import {makeStyles} from "@material-ui/core/styles";
 import Button from "@material-ui/core/Button";
-import Header from "../components/Header/Header";
-import "./sign.css";
-import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth";
-import firebase from "../config/fbConfig";
 
 // Configure FirebaseUI.
 const uiConfig = {
@@ -22,47 +25,63 @@ const uiConfig = {
   ]
 };
 
-const LogInScreen = props => {
-  const [userEmail, setUserEmail] = useState("test@test.com");
-  const [userPassword, setUserPassword] = useState("123456");
+const SignInScreen = props => {
+  const db = useFirestore();
+  const [userEmail, setUserEmail] = useState("");
+  const [userPassword, setUserPassword] = useState("");
+  const [pwConfirm, setPwConfirm] = useState("");
   const [errMsg, setErrMsg] = useState(null);
-  // let errMsg = null;
 
   const handleChange = (e, type) => {
     if (type === "email") {
       setUserEmail(e.target.value);
     } else if (type === "password") {
       setUserPassword(e.target.value);
+    } else if (type === "confirmPw") {
+      setPwConfirm(e.target.value);
     }
   };
   const handleSubmit = e => {
     e.preventDefault();
-    firebase
-      .auth()
-      .signInWithEmailAndPassword(userEmail, userPassword)
-      .then(res => {
-        props.history.push(`/pm/${res.user.uid}`);
-      })
-      .catch(function(error) {
-        if (error.code === "auth/invalid-email") {
-          setErrMsg(<FormattedMessage id="log-in.invalid-email" />);
-        } else if (error.code === "auth/wrong-password") {
-          setErrMsg(<FormattedMessage id="log-in.wrong-password" />);
-        } else if (error.code === "auth/user-not-found") {
-          setErrMsg(<FormattedMessage id="log-in.user-not-found" />);
-        }
-      });
+    if (userPassword === pwConfirm) {
+      firebase
+        .auth()
+        .createUserWithEmailAndPassword(userEmail, userPassword)
+        .then(res => {
+          return writeDbUser(
+            db,
+            res.user.uid,
+            null,
+            "setUserDoc",
+            {createdTime: Date.now()},
+            () => {
+              props.history.push(`/pm/${res.user.uid}`);
+            }
+          );
+        })
+        .catch(error => {
+          if (error.code === "auth/invalid-email") {
+            setErrMsg(<FormattedMessage id="sign-up.invalid-email" />);
+          } else if (error.code === "auth/weak-password") {
+            setErrMsg(<FormattedMessage id="sign-up.weak-password" />);
+          } else if (error.code === "auth/email-already-in-use") {
+            setErrMsg(<FormattedMessage id="sign-up.email-already-in-use" />);
+          }
+        });
+    } else {
+      setErrMsg(<FormattedMessage id="sign-up.password-diff" />);
+    }
   };
 
   return (
     <div className="entry-screen">
       <h2>
-        <FormattedMessage id="log-in.welcome" />
+        <FormattedMessage id="sign-up.welcome" />
       </h2>
       <p>
-        <FormattedMessage id="log-in.instruction" />
+        <FormattedMessage id="sign-up.instruction" />
       </p>
-      <form name="log-in" id="log-in-form" className="entry-form">
+      <form name="sign-up" id="sign-up-form" className="entry-form">
         <TextField
           required
           id="email-input"
@@ -80,24 +99,33 @@ const LogInScreen = props => {
           value={userPassword}
           onChange={e => handleChange(e, "password")}
         />
+        <div className="empty-divider"></div>
+        <TextField
+          required
+          id="confirm-pw-input"
+          label="Confirm Password"
+          type="password"
+          value={pwConfirm}
+          onChange={e => handleChange(e, "confirmPw")}
+        />
         <div className="err-msg">{errMsg}</div>
         <Button
           type="submit"
           variant="contained"
           id="entry-btn"
           onClick={e => handleSubmit(e)}>
-          <FormattedMessage id="log-in.log-in" />
+          <FormattedMessage id="sign-up.create-account" />
         </Button>
       </form>
-      <Link to="/signup">
+      <Link to="/login">
         <div className="switch-link">
-          <FormattedMessage id="log-in.go-to-signup" />
+          <FormattedMessage id="sign-up.go-to-login" />
         </div>
       </Link>
       <div className="entry-divider">
         <div className="divider-line"></div>
         <p>
-          <FormattedMessage id="log-in.3rd-log-in" />
+          <FormattedMessage id="sign-up.3rd-sign-up" />
         </p>
         <div className="divider-line"></div>
       </div>
@@ -106,14 +134,14 @@ const LogInScreen = props => {
     </div>
   );
 };
-const LogIn = props => {
+const SignUp = props => {
   return (
     <Fragment>
       <Header {...props} locale={props.locale} setLocale={props.setLocale} />
       <div className="entry-page">
         <Card className="sign-card">
           <CardContent>
-            <LogInScreen {...props} />
+            <SignInScreen {...props} />
           </CardContent>
         </Card>
       </div>
@@ -128,4 +156,4 @@ let mapStateToProps = (state, props) => {
   };
 };
 
-export default connect(mapStateToProps)(LogIn);
+export default connect(mapStateToProps)(SignUp);
