+++ src/pages/sign.css
@@ -3,7 +3,7 @@
   position: absolute;
   top: 90px;
   width: 100%;
-  height: 100vh; /* Fallback for browsers that do not support Custom Properties */
+  height: 100vh - 90px; /* Fallback for browsers that do not support Custom Properties */
   height: calc(var(--vh, 1vh) * 100 - 90px);
   background-color: rgb(123, 144, 133);
   display: flex;
