+++ src/reducers/sldEditorReducer.js
