+++ src/style.css
@@ -6,14 +6,14 @@
 }
 body {
   font-family: "Source Sans Pro", "Noto Sans TC", sans-serif;
-  overflow-x: hidden;
 }
 
 body,
 #root {
   width: 100vw;
   height: 100vh;
-  overflow: auto;
+  overflow-x: hidden;
+  overflow-y: auto;
 }
 /* Overlay */
